+++ app/src/main/java/org/simple/clinic/registration/name/RegistrationNameViewEffectHandler.kt
@@ -1,0 +1,14 @@
+package org.simple.clinic.registration.name
+
+import org.simple.clinic.mobius.ViewEffectsHandler
+
+class RegistrationNameViewEffectHandler(
+    private val uiActions: RegistrationNameUiActions
+) : ViewEffectsHandler<RegistrationNameViewEffect> {
+  override fun handle(viewEffect: RegistrationNameViewEffect) {
+    when (viewEffect) {
+      is PrefillFields -> uiActions.preFillUserDetails(viewEffect.entry)
+      is ProceedToPinEntry -> uiActions.openRegistrationPinEntryScreen(viewEffect.entry)
+    }
+  }
+}